Refactor: Clean up IBGatewayManager
- Extracted port utilities to separate PortUtils class
- Extracted config management to ConfigUtils class
- Removed unused diagnostic and system resource logging methods
- Removed zombie process detection (overly complex)
- Simplified startup flow by removing redundant environment checks
- Reduced file size from 848 to 505 lines (40% reduction)
- Maintained all public API methods needed by tool handlers
- All builds pass successfully
